Фикс. Правки навигации на main вместо зашлушки menu.
