+++ packages/client/src/constants/router/router.ts
@@ -7,7 +7,6 @@ export enum AppRoutes {
   LEADER_BOARD = '/leaderboard',
   FORUM = '/forum',
   FORUM_TOPIC = '/forum/topic',
-  MENU = '/menu',
   NOT_FOUND = '/not_found',
 }
 
@@ -16,7 +15,6 @@ export const getRouteLogin = () => AppRoutes.LOGIN;
 export const getRouteRegistration = () => AppRoutes.REGISTRATION;
 export const getRouteProfile = (id: string) => `${AppRoutes.PROFILE}/${id}`;
 export const getRouteGame = () => AppRoutes.GAME;
-export const getRouteMenu = () => AppRoutes.MENU;
 export const getRouteLeaderBoard = () => AppRoutes.LEADER_BOARD;
 export const getRouteForum = () => AppRoutes.FORUM;
 export const getRouteForumTopic = (id: string) =>
