--- packages/client/src/pages/menu/index.ts
@@ -1,1 +1,0 @@
-export { MenuPage } from './Menu';
