+++ packages/client/src/widgets/auth-form/AuthForm.tsx
@@ -5,7 +5,7 @@ import { authApi } from '@/services/api/auth/auth-api';
 import styles from './auth-form.module.scss';
 import { Spinner } from '@/shared/components/spinner/Spinner';
 import { useNavigate } from 'react-router-dom';
-import { getRouteMenu, getRouteRegistration } from '@/constants/router/router';
+import { getRouteMain, getRouteRegistration } from '@/constants/router/router';
 import { AuthService } from '@/services/auth/auth';
 import { FormInputText } from '@/shared/components/form-fields/FormInputText';
 import { AuthRequest } from '@/types/auth-request';
@@ -24,7 +24,7 @@ export const AuthForm: React.FC = () => {
     const res = await signin(data);
     if (typeof res === 'string' && res === 'OK') {
       AuthService.setLogged();
-      navigate(getRouteMenu(), {
+      navigate(getRouteMain(), {
         replace: true,
       });
     }
